pyproject.toml
[project]
name = "fastapi-cli-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.119.0",
    "uvicorn[standard]>=0.37.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "aider-install>=0.2.0",
    "requests>=2.0.0",
]
